+++ website.go
@@ -56,8 +56,9 @@ func NewWebsite() *Website {
 			return Wrap(
 				H3(h),
 				Ul(
-					site.AddDrill(h, "-h", "drill/flag_types.go"),
-					site.AddDrill(h, "-h", "drill/flag_names.go"),
+					site.AddDrill(h, "", "drill/flag_types.go"),
+					site.AddDrill(h, "", "drill/flag_names.go"),
+					site.AddDrill(h, "", "drill/cmdline_basic.go"),
 				),
 			)
 		}(),